Punish nodes relatively for oversize blocks
In the case we receive a block that is over our size limit we don't
immediately disconnect. Instead we punish the node based on the amount
of bytes that it was over our limit. (100 points is immediate
disconnect)
This avoids us receiving a block 1 byte over the limit and immediately
banning that node for 24 hours. Instead only ban nodes that have
substantially different ideas about block sizes than we do.