Allow debugging with valgrind
Using valgrind for finding memory leaks, uninitialized variables and
other such issues is awesome, but not posible using the immense amount
of false positives you get from the random engine we currently use.
So here is a simple way to turn on the native random engine to avoid
this.
You need to explicitly compile using that define in your CPPFLAGS;

  export CPPFLAGS=-DUSE_UNSAFE_RANDOM