Fixes in test runner;
Fix the buggy --failed option.
It prints a friendly message if no new or failed tests.
This option is a little dangerous if you have rebuilt in the meantime.
The runner does NOT realize that a previously tested executable might have changed.
Use with care!

Append the username to temp dir created for the logs.
This allows multiple users to use the script on the same system, otherwise it
trips over the unwriteable /tmp dir left by others.

Cleaned up some dead code, comments.