Make logging read its own config file.
This allows a developer or user to configure all the sections to only be
printed above a certain log-level.
It also allows configuring which output channels are to be used.
Currently available are the 'console' and the 'file' output channels.
The console channel has options like printing of the methodname/filename
that the log statement appeared in.