When user has logs.conf, make defaults more sane
This accepts the new reality where logging and debugging are no longer
the same thing, just use the same framework.

If you haev a logs.conf, then the defaults are the same for all
sections, which is that logWarn() will be shown, unless changed in the
config file.
If you don't have this file, the default will be to set the non-global
sections to only show logCritical() and higher.