Add new feature; blocksdatadir config
The operator can add directories to the commandline or config file that
serve as a fallback location for historical block files.

We still look in the ${datapath}/blocks/ first.
Second we iterate through all the values of 'blockdatadir' given and try
to find the blocks there.

For clarity the path passed to 'blockdatadir' should be to a dir that
contains the 'blocks' dir.